[4519] Move create to a file on the API root folder
